feat: add scripts, tests, docs, more scripts and examples. blueprint

+++ test/oracle/OracleSubmissions.t.sol
@@ -23,10 +23,8 @@ contract OracleSubmissionsTest is Test {
         // Prepare operators
         ops = new address[](3);
         pkeys = new uint256[](3);
-        for (uint i = 0; i < 3; i++) {
-            (ops[i], pkeys[i]) = makeAddrAndKey(
-                string(abi.encodePacked("op", i))
-            );
+        for (uint256 i = 0; i < 3; i++) {
+            (ops[i], pkeys[i]) = makeAddrAndKey(string(abi.encodePacked("op", i)));
         }
 
         OracleTypes.FeedConfig memory cfg = OracleTypes.FeedConfig({
@@ -58,11 +56,7 @@ contract OracleSubmissionsTest is Test {
 
         // op1 submits
         _submit(0, roundId, price1);
-        assertEq(
-            oracle.currentRoundId(FEED),
-            roundId,
-            "Round should have started"
-        );
+        assertEq(oracle.currentRoundId(FEED), roundId, "Round should have started");
 
         // op2 submits
         _submit(1, roundId, price2);
@@ -75,8 +69,7 @@ contract OracleSubmissionsTest is Test {
         int256 price3 = 101e8;
         _submit(2, roundId, price3);
 
-        (uint80 latestRoundId2, int256 answer, , uint256 updatedAt, ) = oracle
-            .latestRoundData(FEED);
+        (uint80 latestRoundId2, int256 answer,, uint256 updatedAt,) = oracle.latestRoundData(FEED);
         assertEq(latestRoundId2, roundId, "Round not finalized");
         assertEq(answer, 101e8, "Median incorrect"); // Median of [100, 101, 102]
         assertEq(updatedAt, block.timestamp, "Timestamp wrong");
@@ -89,11 +82,10 @@ contract OracleSubmissionsTest is Test {
         prices[1] = 102e8;
         prices[2] = 101e8;
 
-        PriceLoomOracle.PriceSubmission[]
-            memory subs = new PriceLoomOracle.PriceSubmission[](3);
+        PriceLoomOracle.PriceSubmission[] memory subs = new PriceLoomOracle.PriceSubmission[](3);
         bytes[] memory sigs = new bytes[](3);
 
-        for (uint i = 0; i < 3; i++) {
+        for (uint256 i = 0; i < 3; i++) {
             subs[i] = PriceLoomOracle.PriceSubmission({
                 feedId: FEED,
                 roundId: roundId,
@@ -107,8 +99,7 @@ contract OracleSubmissionsTest is Test {
 
         oracle.submitSignedBatch(FEED, subs, sigs);
 
-        (uint80 latestRoundId, int256 answer, , uint256 updatedAt, ) = oracle
-            .latestRoundData(FEED);
+        (uint80 latestRoundId, int256 answer,, uint256 updatedAt,) = oracle.latestRoundData(FEED);
         assertEq(latestRoundId, roundId, "Round not finalized");
         assertEq(answer, 101e8, "Median incorrect"); // Median of [100, 101, 102]
         assertEq(updatedAt, block.timestamp, "Timestamp wrong");
@@ -135,8 +126,7 @@ contract OracleSubmissionsTest is Test {
         // Now, anyone can trigger the finalization
         oracle.poke(FEED);
 
-        (uint80 latestRoundId2, int256 answer, , uint256 updatedAt, ) = oracle
-            .latestRoundData(FEED);
+        (uint80 latestRoundId2, int256 answer,, uint256 updatedAt,) = oracle.latestRoundData(FEED);
         assertEq(latestRoundId2, roundId, "Round not finalized after timeout");
         assertEq(answer, 101e8, "Median incorrect"); // Median of [100, 102]
         assertEq(updatedAt, block.timestamp, "Timestamp wrong");
@@ -149,9 +139,7 @@ contract OracleSubmissionsTest is Test {
         _submit(1, roundId1, 101e8);
         _submit(2, roundId1, 102e8);
 
-        (uint80 latestRoundId1, int256 answer1, , , ) = oracle.latestRoundData(
-            FEED
-        );
+        (uint80 latestRoundId1, int256 answer1,,,) = oracle.latestRoundData(FEED);
         assertEq(latestRoundId1, 1, "Round 1 not finalized");
         assertEq(answer1, 101e8, "Median for round 1 incorrect");
 
@@ -165,13 +153,8 @@ contract OracleSubmissionsTest is Test {
         // Poke the oracle to handle the timeout
         oracle.poke(FEED);
 
-        (
-            uint80 latestRoundId2,
-            int256 answer2,
-            ,
-            uint256 updatedAt,
-            uint80 answeredInRound
-        ) = oracle.latestRoundData(FEED);
+        (uint80 latestRoundId2, int256 answer2,, uint256 updatedAt, uint80 answeredInRound) =
+            oracle.latestRoundData(FEED);
         assertEq(latestRoundId2, 2, "Round 2 not finalized");
         assertEq(answer2, 101e8, "Price should be forwarded from round 1");
         assertEq(answeredInRound, 1, "answeredInRound should be 1");
@@ -182,13 +165,12 @@ contract OracleSubmissionsTest is Test {
         uint80 roundId = 2; // Round 1 is not started yet
         int256 price = 100e8;
 
-        PriceLoomOracle.PriceSubmission memory sub = PriceLoomOracle
-            .PriceSubmission({
-                feedId: FEED,
-                roundId: roundId,
-                answer: price,
-                validUntil: block.timestamp + 60
-            });
+        PriceLoomOracle.PriceSubmission memory sub = PriceLoomOracle.PriceSubmission({
+            feedId: FEED,
+            roundId: roundId,
+            answer: price,
+            validUntil: block.timestamp + 60
+        });
         _submitExpectRevert(0, sub, WrongRound.selector);
     }
 
@@ -196,13 +178,12 @@ contract OracleSubmissionsTest is Test {
         uint80 roundId = 1;
         int256 price = 100e8;
 
-        PriceLoomOracle.PriceSubmission memory sub = PriceLoomOracle
-            .PriceSubmission({
-                feedId: FEED,
-                roundId: roundId,
-                answer: price,
-                validUntil: block.timestamp - 1
-            });
+        PriceLoomOracle.PriceSubmission memory sub = PriceLoomOracle.PriceSubmission({
+            feedId: FEED,
+            roundId: roundId,
+            answer: price,
+            validUntil: block.timestamp - 1
+        });
 
         bytes32 digest = oracle.getTypedDataHash(sub);
         (uint8 v, bytes32 r, bytes32 s) = vm.sign(pkeys[0], digest);
@@ -216,13 +197,12 @@ contract OracleSubmissionsTest is Test {
         uint80 roundId = 1;
         int256 price = 2e20; // > maxPrice
 
-        PriceLoomOracle.PriceSubmission memory sub = PriceLoomOracle
-            .PriceSubmission({
-                feedId: FEED,
-                roundId: roundId,
-                answer: price,
-                validUntil: block.timestamp + 60
-            });
+        PriceLoomOracle.PriceSubmission memory sub = PriceLoomOracle.PriceSubmission({
+            feedId: FEED,
+            roundId: roundId,
+            answer: price,
+            validUntil: block.timestamp + 60
+        });
         _submitExpectRevert(0, sub, OutOfBounds.selector);
     }
 
@@ -232,21 +212,16 @@ contract OracleSubmissionsTest is Test {
 
         _submit(0, roundId, price);
 
-        PriceLoomOracle.PriceSubmission memory sub = PriceLoomOracle
-            .PriceSubmission({
-                feedId: FEED,
-                roundId: roundId,
-                answer: price,
-                validUntil: block.timestamp + 60
-            });
+        PriceLoomOracle.PriceSubmission memory sub = PriceLoomOracle.PriceSubmission({
+            feedId: FEED,
+            roundId: roundId,
+            answer: price,
+            validUntil: block.timestamp + 60
+        });
         _submitExpectRevert(0, sub, DuplicateSubmission.selector);
     }
 
-    function _submitExpectRevert(
-        uint opIdx,
-        PriceLoomOracle.PriceSubmission memory sub,
-        bytes4 selector
-    ) internal {
+    function _submitExpectRevert(uint256 opIdx, PriceLoomOracle.PriceSubmission memory sub, bytes4 selector) internal {
         bytes32 digest = oracle.getTypedDataHash(sub);
         (uint8 v, bytes32 r, bytes32 s) = vm.sign(pkeys[opIdx], digest);
         bytes memory sig = abi.encodePacked(r, s, v);
@@ -254,14 +229,13 @@ contract OracleSubmissionsTest is Test {
         oracle.submitSigned(FEED, sub, sig);
     }
 
-    function _submit(uint opIdx, uint80 roundId, int256 answer) internal {
-        PriceLoomOracle.PriceSubmission memory sub = PriceLoomOracle
-            .PriceSubmission({
-                feedId: FEED,
-                roundId: roundId,
-                answer: answer,
-                validUntil: block.timestamp + 60
-            });
+    function _submit(uint256 opIdx, uint80 roundId, int256 answer) internal {
+        PriceLoomOracle.PriceSubmission memory sub = PriceLoomOracle.PriceSubmission({
+            feedId: FEED,
+            roundId: roundId,
+            answer: answer,
+            validUntil: block.timestamp + 60
+        });
 
         bytes32 digest = oracle.getTypedDataHash(sub);
         (uint8 v, bytes32 r, bytes32 s) = vm.sign(pkeys[opIdx], digest);
